fix beta as double

+++ sw/blas/gemm/data/datagen.py
@@ -102,7 +102,7 @@ def emit_header(**kwargs):
     data_str += [format_scalar_definition('uint32_t', 'K', kwargs['K'])]
     data_str += [format_scalar_definition('uint32_t', 'TA', int(kwargs['ta']))]
     data_str += [format_scalar_definition('uint32_t', 'TB', int(kwargs['tb']))]
-    data_str += [format_scalar_definition('uint32_t', 'BETA', kwargs['beta'])]
+    data_str += [format_scalar_definition('double', 'BETA', kwargs['beta'])]
     data_str += [format_scalar_definition('uint32_t', 'dtype_size', kwargs['prec']//8)]
     data_str += [format_scalar_definition('uint32_t', 'expand', kwargs['expand'])]
     data_str += [format_vector_definition(C_TYPES[str(kwargs['prec'])], 'a', a.flatten(),
